backend/requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
sqlalchemy
mercadopago
httpx
pydantic>=2.0
pydantic-settings>=2.0
pytest
python-dateutil
